reducer の処理を他のファイルに分けた
